ipc: convert ipcs_idr to XArray
The XArray has better loops than the IDR has, removing the need to
open-code them.  We also don't need to call idr_destroy() any more. 
Allocating the ID is a little tricky due to needing to get 'seq' correct. 
Open-code a variant of __xa_alloc() which lets us set the ID and the seq
before depositing the pointer in the array.

Link: http://lkml.kernel.org/r/20200326151418.27545-1-willy@infradead.org
Signed-off-by: Matthew Wilcox <willy@infradead.org>
Cc: Manfred Spraul <manfred@colorfullife.com>
Cc: Davidlohr Bueso <dave@stgolabs.net>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>