umh: fix processed error when UMH_WAIT_PROC is used
When UMH_WAIT_PROC is used we call kernel_wait4().  This is the *only*
place in the kernel where we actually inspect the error code.  Prior to
this patch we returned the value from the wait call, and that technically
requires us to use wrappers such as WEXITSTATUS().  We either fix all
callers to start using WEXITSTATUS() and friends *or* we do address this
within the umh code and let the callers get the actual error code.

The way we use kernel_wait4() on the umh is with the options set to 0, and
when this is done the wait call only waits for terminated children. 
Because of this, there is no point to complicate checks for the umh with
W*() calls.  That would make the checks complex, redundant, and simply not
needed.

By making the umh do the checks for us we keep users kernel_wait4() at
bay, and promote avoiding introduction of further W*() macros and the
complexities this can bring.

There were only a few callers which properly checked for the error status
using open-coded solutions.  We remove them as they are no longer needed,
and also remove open coded implicit uses of W*() uses which should never
trigger given that the options passed to wait is 0.

The only helpers we really need are for termination, so we just include
those, and we prefix our W*() helpers with K.

Since all this does is *correct* an error code, if one was found, this
change only fixes reporting the *correct* error, and there are two places
where this matters, and which this patch fixes:

  * request_module() used to fail with an error code of
    256 when a module was not found. Now it properly
    returns 1.

  * fs/nfsd/nfs4recover.c: we never were disabling the
    upcall as the error code of -ENOENT or -EACCES was
    *never* properly checked for.

Link: http://lkml.kernel.org/r/20200610154923.27510-5-mcgrof@kernel.org
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>
Reported-by: Tiezhu Yang <yangtiezhu@loongson.cn>
Cc: Alexei Starovoitov <ast@kernel.org>
Cc: Al Viro <viro@zeniv.linux.org.uk>
Cc: Christian Brauner <christian.brauner@ubuntu.com>
Cc: Chuck Lever <chuck.lever@oracle.com>
Cc: David Howells <dhowells@redhat.com>
Cc: David S. Miller <davem@davemloft.net>
Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Cc: Jakub Kicinski <kuba@kernel.org>
Cc: James Morris <jmorris@namei.org>
Cc: Jarkko Sakkinen <jarkko.sakkinen@linux.intel.com>
Cc: J. Bruce Fields <bfields@fieldses.org>
Cc: Jens Axboe <axboe@kernel.dk>
Cc: Josh Triplett <josh@joshtriplett.org>
Cc: Kees Cook <keescook@chromium.org>
Cc: Lars Ellenberg <lars.ellenberg@linbit.com>
Cc: Nikolay Aleksandrov <nikolay@cumulusnetworks.com>
Cc: Philipp Reisner <philipp.reisner@linbit.com>
Cc: Roopa Prabhu <roopa@cumulusnetworks.com>
Cc: "Serge E. Hallyn" <serge@hallyn.com>
Cc: Sergei Trofimovich <slyfox@gentoo.org>
Cc: Sergey Kvachonok <ravenexp@gmail.com>
Cc: Shuah Khan <shuah@kernel.org>
Cc: Tony Vroon <chainsaw@gentoo.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>